Move extension class to right folder SN-445
